Remove it00 (put into commentary)
I removed it00 from the global datasets (I put it in a commentary) and the net datasets since it has no value for PIT (hxit). In the end, there are no obs left and it makes the program stop after ie00 since stata cannot display anything.
I did not notice this problem before (@michaelzemmour did) so I did not test the program for subsequent countryyears (lu00 and mexico).