Add new years to the database + fix on pensions
I add several new years to the dataset to have the most recent point for each country
In addition, I automatize the previous "fixpension_datasets1" in the new missing_values program. Note that this program could be adapted to other variables than pensions (e.g. other types of income, etc.): the general idea (as done before in the fixpension_ datasets1 program) is to reconstitute a variable with the value of its subvariables when the value of the aggregate is missing and the value of the subvariables is not.