Impute PSSCEE for Italy
In 2014, no hxits for Italy in the data so we need to impute them.

In turn, hxit is well defined and its mean is comparable to hxit in 2010 (it's a bit lower but the mean of hxits in 2010 makes it clear that the difference in hxit is not due to the absence of hxits). Therefore, I think we can define hxiti as hxit - hxits if hxiti==.: it shouldn't change anything in global redistribution analysis (because we only use hxit) but it can give additional information. I add this in the general program (define_tax_transfers) because it can be applied to other countries as well